pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "auth-solid"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "async-sqlalchemy>=1.0.0",
    "bcrypt>=4.2.1",
    "dishka>=1.4.2",
    "fastapi[standard]>=0.115.6",
    "jwt>=1.3.1",
    "uuid6>=2024.7.10",
]
